extract test_certstor_utils
